Switch to using fprintf in ignition Console.
fprintf() is considered to be an atomic output, while
using the stream operator (<<) is not.  Thus, using
the stream operator inside of the Console was sometimes
causing control characters to be put on different lines
than the rest of the string, leading to really ugly output.
By forming up the string into a stringstream and then using
fprintf(), we get rid of that particular issue.  Note that
this only improves, but doesn't completely solve the issue;
if the user of ignition::common::Console uses the stream
operator to concatenate individual strings, the same problem
can happen, but at least the control codes will not be
everywhere.