Fix IGN_HOMEDIR on Windows.
HOMEPATH alone is a pretty wicked concept. For its use to make sense, it has to be joined with HOMEDRIVE, because HOMEPATH is a path relative to HOMEDRIVE. So when you lookup just env("HOMEPATH") and use it as a file path, it gets resolved relative to any random drive letter that currently happens to be in CWD. I guess this could not be the inteded behavior.

USERPROFILE, on the other hand, contains the full absolute path to the user's home directory.

Signed-off-by: Martin Pecka <peckama2@fel.cvut.cz>