Fix loading collada files with multiple texcoord sets using the same offset (#208)
When loading texture coordinates, there was an assumption in the ColladaLoader that the offset values (which specify the index in which the uv coordinates should be read from) need to be unique. This is found that be incorrect assumption after looking at Collada 1.5 spec (link to pdf. There is an example on page 89).