Fix severity level of gzlog (#635)
The severity level of `gzlog` was incorrectly set to `err`. Changing this to `trace` solves two problems:
1. The log message is now a lot less alarming 
2. When running `gz sim -v4`, this message is not shown on the console. It will only be logged to file. It does show up on the console if you run with `-v5`. 
 
The second point is technically a behavior change since in Harmonic, regardless of the verbosity level, `gzlog` would never show up on the console. However, duplicating that behavior with the new Console implementation proved to be very complicated. So, I propose we change the meaning of `gzlog` to be equivalent to `gztrace`.

Signed-off-by: Addisu Z. Taddese <addisu@openrobotics.org>